feat(dev-server-core): always decude mimetype in serve hook
