+++ packages/dev-server-core/package.json
@@ -27,7 +27,11 @@
     "test:watch": "mocha 'test/**/*.test.ts' --recursive --require ts-node/register --reporter progress --watch --watch-files src,test/tests"
   },
   "files": [
-    "dist"
+    ".self-signed-dev-server-ssl.cert",
+    ".self-signed-dev-server-ssl.key",
+    "dist",
+    "test-helpers.d.ts",
+    "test-helpers.js"
   ],
   "keywords": [
     "web",
