+++ packages/dev-server-core/src/index.ts
@@ -6,3 +6,6 @@ export { getRequestBrowserPath } from './utils';
 export { getRequestFilePath } from './utils';
 export { Logger } from './logger/Logger';
 export { EventStreamManager } from './event-stream/EventStreamManager';
+export { FSWatcher } from 'chokidar';
+export { default as Koa, Context } from 'koa';
+export { Server } from 'net';
