+++ packages/dev-server-core/test-helpers.d.ts
@@ -1,0 +1,1 @@
+export * from './dist/test-helpers';
