+++ packages/dev-server-core/test-helpers.js
@@ -1,0 +1,2 @@
+/* eslint-disable */
+module.exports = require('./dist/test-helpers');
