+++ packages/dev-server-core/test/tests/helpers.ts
@@ -1,38 +1,17 @@
-import portfinder from 'portfinder';
 import path from 'path';
-import { DevServer } from '../../src/server/DevServer';
+import {
+  createTestServer as originalCreateTestServer,
+  timeout,
+  fetchText,
+  expectIncludes,
+} from '../../src/test-helpers';
 import { Config } from '../../src/Config';
 
-const defaultConfig: Omit<Config, 'port'> = {
-  rootDir: path.resolve(__dirname, '..', 'fixtures', 'basic'),
-  hostname: 'localhost',
-  middleware: [],
-  plugins: [],
-};
-
-const mockLogger = {
-  log() {
-    //
-  },
-  debug() {
-    //
-  },
-  error() {
-    //
-  },
-  warn() {
-    //
-  },
-  logSyntaxError() {
-    //
-  },
-};
-
-export async function createTestServer(config?: Partial<Config>) {
-  const port = await portfinder.getPortPromise();
-  const server = new DevServer({ ...defaultConfig, ...config, port }, mockLogger);
-  await server.start();
-  return { server, port, host: `http://localhost:${port}` };
+export function createTestServer(config: Partial<Config> = {}) {
+  return originalCreateTestServer({
+    ...config,
+    rootDir: path.resolve(__dirname, '..', 'fixtures', 'basic'),
+  });
 }
 
-export const timeout = (ms = 0) => new Promise(resolve => setTimeout(resolve, ms));
+export { timeout, fetchText, expectIncludes };
